Shared hook for fetching proof
